+++ StartSch/Program.cs
@@ -27,13 +27,13 @@ using StartSch.Services;
 var builder = WebApplication.CreateBuilder(args);
 
 // Modules
-// builder.Services.AddModule<CmschModule>();
-// builder.Services.AddModule<GeneralEventModule>();
-// builder.Services.AddModule<KthBmeHuModule>();
-// builder.Services.AddModule<SchBodyModule>();
-// builder.Services.AddModule<SchPincerModule>();
-// builder.Services.AddModule<VikBmeHuModule>();
-// builder.Services.AddModule<VikHkModule>();
+builder.Services.AddModule<CmschModule>();
+builder.Services.AddModule<GeneralEventModule>();
+builder.Services.AddModule<KthBmeHuModule>();
+builder.Services.AddModule<SchBodyModule>();
+builder.Services.AddModule<SchPincerModule>();
+builder.Services.AddModule<VikBmeHuModule>();
+builder.Services.AddModule<VikHkModule>();
 
 // Services
 builder.Services.AddSingletonAndHostedService<BackgroundTaskManager>();
